remove pro feature from services
