+++ Settings-default.toml
@@ -144,11 +144,6 @@ allowed_drivers = [
 # Whether to allow requests to `/anonymous` that return a valid session.
 anonymous_access = true
 
-# Specifies a fixed session token. Otherwise, a random token is generated.
-# This can be directly used for Bearer authentication in HTTP requests.
-# This setting is ignored in the Pro version.
-# fixed_session_token = "18fec623-6600-41af-b82b-24ccf47cb9f9"
-
 # Settings for Geo Engine Pro
 [user]
 registration = true
