+++ services/src/api/mod.rs
@@ -1,4 +1,3 @@
-#[cfg(not(feature = "pro"))] // TODO: this needs to be done without mentioning pro
 pub mod apidoc;
 pub mod cli;
 pub mod handlers;
