+++ services/src/lib.rs
@@ -5,7 +5,6 @@ pub mod error;
 pub mod layers;
 pub mod machine_learning;
 pub mod projects;
-#[cfg(not(feature = "pro"))]
 pub mod server;
 pub mod stac;
 #[macro_use]
@@ -13,8 +12,7 @@ pub mod util;
 pub mod tasks;
 pub mod workflows;
 
-/// Compiles Geo Engine Pro
-#[cfg(feature = "pro")]
+/// TODO: merge with rest of the code
 pub mod pro;
 
 pub use geoengine_datatypes::test_data;
