+++ services/src/server.rs
@@ -1,73 +1,28 @@
-use crate::api::apidoc::ApiDoc;
 use crate::api::handlers;
-use crate::contexts::{PostgresContext, SimpleApplicationContext};
+use crate::contexts::{ApplicationContext, SessionContext};
 use crate::error::{Error, Result};
-use crate::util::config;
-use crate::util::config::get_config_element;
+use crate::pro;
+use crate::pro::api::ApiDoc;
+use crate::pro::contexts::ProPostgresContext;
+use crate::util::config::{self, get_config_element};
 use crate::util::middleware::OutputRequestId;
 use crate::util::server::{
     calculate_max_blocking_threads_per_worker, configure_extractors, connection_init,
     log_server_info, render_404, render_405, serve_openapi_json, CustomRootSpanBuilder,
 };
 use actix_files::Files;
-use actix_web::{http, middleware, web, App, HttpServer};
+use actix_web::{http, middleware, web, App, FromRequest, HttpServer};
+use bb8_postgres::tokio_postgres::NoTls;
+use geoengine_datatypes::raster::TilingSpecification;
+use geoengine_operators::engine::ChunkByteSize;
 use geoengine_operators::util::gdal::register_gdal_drivers_from_list;
 use log::info;
+use pro::contexts::{ProApplicationContext, ProGeoEngineDb};
 use std::net::SocketAddr;
 use std::path::PathBuf;
 use tracing_actix_web::TracingLogger;
 use utoipa::OpenApi;
 use utoipa_swagger_ui::SwaggerUi;
-
-/// Starts the webserver for the Geo Engine API.
-///
-/// # Panics
-///  * may panic if the `Postgres` backend is chosen without compiling the `postgres` feature
-///
-///
-pub async fn start_server(static_files_dir: Option<PathBuf>) -> Result<()> {
-    log_server_info()?;
-
-    let web_config: crate::util::config::Web = get_config_element()?;
-    let session_config: crate::util::config::Session = get_config_element()?;
-
-    if let Some(session_token) = session_config.fixed_session_token {
-        info!("Fixed Session Token: {session_token}");
-    }
-
-    let data_path_config: config::DataProvider = get_config_element()?;
-
-    let chunk_byte_size = config::get_config_element::<config::QueryContext>()?
-        .chunk_byte_size
-        .into();
-
-    let tiling_spec = config::get_config_element::<config::TilingSpecification>()?.into();
-
-    register_gdal_drivers_from_list(config::get_config_element::<config::Gdal>()?.allowed_drivers);
-
-    let db_config = config::get_config_element::<config::Postgres>()?;
-
-    let ctx = PostgresContext::new_with_data(
-        bb8_postgres::tokio_postgres::Config::try_from(db_config)?,
-        tokio_postgres::NoTls,
-        data_path_config.dataset_defs_path,
-        data_path_config.provider_defs_path,
-        data_path_config.layer_defs_path,
-        data_path_config.layer_collection_defs_path,
-        tiling_spec,
-        chunk_byte_size,
-    )
-    .await?;
-
-    start(
-        static_files_dir,
-        web_config.bind_address,
-        web_config.api_prefix,
-        web_config.version_api,
-        ctx,
-    )
-    .await
-}
 
 async fn start<C>(
     static_files_dir: Option<PathBuf>,
@@ -77,7 +32,9 @@ async fn start<C>(
     app_ctx: C,
 ) -> Result<(), Error>
 where
-    C: SimpleApplicationContext,
+    C: ProApplicationContext,
+    C::Session: FromRequest,
+    <<C as ApplicationContext>::SessionContext as SessionContext>::GeoEngineDB: ProGeoEngineDb,
 {
     let wrapped_ctx = web::Data::new(app_ctx);
 
@@ -86,11 +43,12 @@ where
     HttpServer::new(move || {
         let mut api = web::scope(&api_prefix)
             .configure(configure_extractors)
-            .configure(handlers::datasets::init_dataset_routes::<C>)
+            .configure(pro::api::handlers::datasets::init_dataset_routes::<C>)
             .configure(handlers::layers::init_layer_routes::<C>)
+            .configure(pro::api::handlers::permissions::init_permissions_routes::<C>)
             .configure(handlers::plots::init_plot_routes::<C>)
             .configure(handlers::projects::init_project_routes::<C>)
-            .configure(handlers::session::init_session_routes::<C>)
+            .configure(pro::api::handlers::users::init_user_routes::<C>)
             .configure(handlers::spatial_references::init_spatial_reference_routes::<C>)
             .configure(handlers::upload::init_upload_routes::<C>)
             .configure(handlers::tasks::init_task_routes::<C>)
@@ -98,6 +56,7 @@ where
             .configure(handlers::wfs::init_wfs_routes::<C>)
             .configure(handlers::wms::init_wms_routes::<C>)
             .configure(handlers::workflows::init_workflow_routes::<C>)
+            .configure(pro::api::handlers::machine_learning::init_ml_routes::<C>)
             .route(
                 "/available",
                 web::get().to(crate::util::server::available_handler),
@@ -108,7 +67,7 @@ where
         api = serve_openapi_json(
             api,
             &mut api_urls,
-            "Geo Engine",
+            "Geo Engine Pro",
             "../api-docs/openapi.json",
             "/api-docs/openapi.json",
             openapi.clone(),
@@ -158,4 +117,110 @@ where
     .run()
     .await
     .map_err(Into::into)
+}
+
+///
+///  Starts the webserver for the Geo Engine API.
+///
+pub async fn start_server(static_files_dir: Option<PathBuf>) -> Result<()> {
+    log_server_info()?;
+
+    let user_config: crate::pro::util::config::User = get_config_element()?;
+    let oidc_config: crate::pro::util::config::Oidc = get_config_element()?;
+    let web_config: crate::util::config::Web = get_config_element()?;
+    let open_telemetry: crate::pro::util::config::OpenTelemetry = get_config_element()?;
+    let cache_config: crate::pro::util::config::Cache = get_config_element()?;
+    let quota_config: crate::pro::util::config::Quota = get_config_element()?;
+
+    if user_config.registration {
+        info!("User Registration: enabled");
+    } else {
+        info!("User Registration: disabled");
+    }
+
+    if oidc_config.enabled {
+        info!("OIDC: enabled");
+    } else {
+        info!("OIDC: disabled");
+    }
+
+    if open_telemetry.enabled {
+        info!("OpenTelemetry Tracing: enabled");
+        info!(
+            "OpenTelemetry Tracing Endpoint: {}",
+            open_telemetry.endpoint
+        );
+    } else {
+        info!("OpenTelemetry Tracing: disabled");
+    }
+
+    if cache_config.enabled {
+        info!("Cache: enabled ({} MB)", cache_config.size_in_mb);
+    } else {
+        info!("Cache: disabled");
+    }
+
+    info!("QuotaTracking: {:?}", quota_config.mode);
+
+    let data_path_config: config::DataProvider = get_config_element()?;
+
+    let chunk_byte_size = config::get_config_element::<config::QueryContext>()?
+        .chunk_byte_size
+        .into();
+
+    let tiling_spec = config::get_config_element::<config::TilingSpecification>()?.into();
+
+    register_gdal_drivers_from_list(config::get_config_element::<config::Gdal>()?.allowed_drivers);
+
+    start_postgres(
+        data_path_config,
+        tiling_spec,
+        oidc_config,
+        chunk_byte_size,
+        static_files_dir,
+        web_config,
+        cache_config,
+        quota_config,
+    )
+    .await
+}
+
+#[allow(clippy::too_many_arguments)]
+async fn start_postgres(
+    data_path_config: config::DataProvider,
+    tiling_spec: TilingSpecification,
+    oidc_config: crate::pro::util::config::Oidc,
+    chunk_byte_size: ChunkByteSize,
+    static_files_dir: Option<PathBuf>,
+    web_config: config::Web,
+    cache_config: crate::pro::util::config::Cache,
+    quota_config: crate::pro::util::config::Quota,
+) -> Result<()> {
+    {
+        let db_config = config::get_config_element::<config::Postgres>()?;
+
+        let ctx = ProPostgresContext::new_with_data(
+            bb8_postgres::tokio_postgres::Config::try_from(db_config)?,
+            NoTls,
+            data_path_config.dataset_defs_path,
+            data_path_config.provider_defs_path,
+            data_path_config.layer_defs_path,
+            data_path_config.layer_collection_defs_path,
+            tiling_spec,
+            chunk_byte_size,
+            oidc_config,
+            cache_config,
+            quota_config,
+        )
+        .await?;
+
+        start(
+            static_files_dir,
+            web_config.bind_address,
+            web_config.api_prefix,
+            web_config.version_api,
+            ctx,
+        )
+        .await
+    }
 }
